+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Added `Content.blank` https://github.com/Textualize/textual/pull/6264
+
 ## [6.7.1] - 2025-12-1
 
 ### Fixed